Add files via upload
You are given an array of arrays of integers, where each array corresponds to a row in a triangle of numbers. We define a path in the triangle to start at the top and go down one row at a time to an adjacent value, eventually ending with an entry on the bottom row. Write a program that returns the weight of the maximum weight path.