Add files via upload
Given an array of integers and a number k, where 1 <= k <= length of the array, compute the maximum values of each subarray of length k. Do this in O(n) time and O(k) space. You can modify the input array in-place and you do not need to store the results. You can simply print them out as you compute them